requirements.txt
fastapi[standard]
uvicorn
pydantic
sqlalchemy
sqlalchemy-utils
sqlmodel
pymysql
httpx
python-dotenv
pillow
numpy
apscheduler
pytest
mysqlclient
cryptography
requests